Moved KVSConfig.java out of atlas-spi to atlas-api
