+++ atlasdb-api/src/main/java/com/palantir/atlasdb/keyvalue/api/KeyValueServiceConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.atlasdb.spi;
+package com.palantir.atlasdb.keyvalue.api;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
