+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/cassandra/CassandraAtlasDbFactory.java
@@ -20,19 +20,19 @@ import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.cassandra.CassandraKeyValueService;
 import com.palantir.atlasdb.keyvalue.cassandra.CassandraTimestampBoundStore;
 import com.palantir.atlasdb.spi.AtlasDbFactory;
-import com.palantir.atlasdb.spi.KeyValueServiceConfig;
+import com.palantir.atlasdb.keyvalue.api.KeyValueServiceConfig;
 import com.palantir.timestamp.PersistentTimestampService;
 import com.palantir.timestamp.TimestampService;
 
 public class CassandraAtlasDbFactory implements AtlasDbFactory {
-    
+
     @Override
     public KeyValueService createRawKeyValueService(KeyValueServiceConfig config) {
         Preconditions.checkArgument(config instanceof CassandraKeyValueServiceConfig,
                 "CassandraAtlasDbFactory expects a configuration of type CassandraKeyValueServiceConfig, found %s", config.getClass());
         return createKv((CassandraKeyValueServiceConfig) config);
     }
-    
+
     private static CassandraKeyValueService createKv(CassandraKeyValueServiceConfig config) {
         return CassandraKeyValueService.create(CassandraKeyValueServiceConfigManager.createSimpleManager(config));
     }
@@ -43,7 +43,7 @@ public class CassandraAtlasDbFactory implements AtlasDbFactory {
                 "TimestampService must be created from an instance of CassandraKeyValueService, found %s", rawKvs.getClass());
         return PersistentTimestampService.create(CassandraTimestampBoundStore.create((CassandraKeyValueService) rawKvs));
     }
-    
+
     @Override
     public String getType() {
         return CassandraKeyValueServiceConfig.TYPE;
