+++ atlasdb-cassandra/src/main/resources/META-INF/services/com.palantir.atlasdb.keyvalue.api.KeyValueServiceConfig
