+++ atlasdb-config/src/main/resources/META-INF/services/io.dropwizard.jackson.Discoverable
@@ -1,1 +1,1 @@
-com.palantir.atlasdb.spi.KeyValueServiceConfig
+com.palantir.atlasdb.keyvalue.api.KeyValueServiceConfig
