+++ atlasdb-rocksdb/src/main/java/com/palantir/atlasdb/rocksdb/RocksDbKeyValueServiceConfig.java
@@ -28,7 +28,7 @@ import com.google.common.base.Preconditions;
 import com.palantir.atlasdb.keyvalue.rocksdb.impl.ImmutableWriteOpts;
 import com.palantir.atlasdb.keyvalue.rocksdb.impl.RocksComparatorName;
 import com.palantir.atlasdb.keyvalue.rocksdb.impl.WriteOpts;
-import com.palantir.atlasdb.spi.KeyValueServiceConfig;
+import com.palantir.atlasdb.keyvalue.api.KeyValueServiceConfig;
 
 @JsonDeserialize(as = ImmutableRocksDbKeyValueServiceConfig.class)
 @JsonSerialize(as = ImmutableRocksDbKeyValueServiceConfig.class)
